+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next [????.??.??]
+* Allow the test suite to compile with GHC 8.8.
+
 ## 3.8 [2019.03.19]
 * Remove the `TextShow I16` instance from `TextShow.Data.Text`. The usefulness
   of this instance was dubious at best, and moreover, it was actively causing