+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### next [????.??.??]
+### 3.8.2 [2019.05.02]
 * Make the `TextShow` instances for `UArray` and `Fixed` use the correct
   precedence on `base-4.13` or later.
 * Support deriving `TextShow(1)(2)` instances for data types with fields